+++ R/cfacutoff.h.R
@@ -44,7 +44,8 @@ cfacutoffOptions <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
                 "iterations",
                 iterations,
                 default=250,
-                min=50)
+                min=50,
+                max=5000)
             private$..seed <- jmvcore::OptionInteger$new(
                 "seed",
                 seed,
@@ -125,7 +126,7 @@ cfacutoffResults <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
                 options=options,
                 name="loadingsTable",
                 title="Standardized Loadings vs Simulated Expected Range",
-                rows=0,
+                rows="(vars)",
                 clearWith=list(
                     "vars",
                     "estimator",