+++ R/personfit.h.R
@@ -59,7 +59,8 @@ personfitOptions <- if (requireNamespace("jmvcore", quietly=TRUE)) R6::R6Class(
                 "iterations",
                 iterations,
                 default=500,
-                min=100)
+                min=100,
+                max=5000)
             private$..seed <- jmvcore::OptionInteger$new(
                 "seed",
                 seed,